requirements.txt
Django>=4.2
django-ckeditor
Pillow
kavenegar
python-dotenv
